+++ CHANGELOG.md
@@ -10,6 +10,10 @@
 
 ### IMPROVEMENTS
 
+- `[state]` use fast validator loading for `BlockExecutor` commit-info paths,
+  avoiding unnecessary proposer-priority reconstruction on cache misses
+  ([\#6039](https://github.com/cometbft/cometbft/pull/6039))
+
 ### FEATURES
 
 ### STATE-BREAKING